refactor: add latest photo of Adonai Vera
